# crates/api/src/routes/summary/tests.rs
use std::sync::Arc;

use axum::body::Body;
use axum::http::{Request, StatusCode};
use http_body_util::BodyExt;
use store::{
    ContainerDiskSample, ContainerNetworkSample, ContainerSample, ContainerStatusSample,
    DiskSample, HostStatusRow, MemRow, Store,
};
use tower::ServiceExt;

use crate::AppState;

fn mem_row(time: i64, used: u64) -> MemRow {
    MemRow {
        time,
        total: 100,
        available: 100 - used,
        used,
        used_percent: used as f64,
        free: 100 - used,
    }
}

fn disk(mount: &str, used: u64) -> DiskSample {
    DiskSample {
        mount: mount.into(),
        total: 100,
        used,
        available: 100 - used,
        used_percent: used as f64,
    }
}

fn container(id: &str, cpu: f64, mem_used: u64) -> ContainerSample {
    ContainerSample {
        container_id: id.into(),
        cpu_percent: cpu,
        mem_total: 100,
        mem_available: 100 - mem_used,
        mem_used,
        mem_used_percent: mem_used as f64,
        mem_free: 100 - mem_used,
    }
}

fn container_disk(id: &str, writable: u64) -> ContainerDiskSample {
    ContainerDiskSample {
        container_id: id.into(),
        writable_layer: writable,
        volumes_total: writable * 2,
    }
}

fn state(store: Store) -> Arc<AppState> {
    let mut config = config::Config::load_for_test();
    config.token = "secret".into();
    Arc::new(AppState {
        auth_header: format!("Bearer {}", config.token),
        config: Arc::new(config),
        store,
        sampler: Arc::new(tokio::sync::Mutex::new(collector::HostSampler::new())),
        memory: Arc::new(crate::CachedMemory::new(
            collector::HostSampler::new().sample_memory(),
        )),
        history_queries: Arc::new(tokio::sync::Semaphore::new(
            crate::MAX_CONCURRENT_HISTORY_QUERIES,
        )),
        analytics_queries: Arc::new(tokio::sync::Semaphore::new(
            crate::MAX_CONCURRENT_ANALYTICS_QUERIES,
        )),
        analytics: None,
        geoip_attribution: Arc::new(std::sync::RwLock::new(None)),
        push_status: Default::default(),
    })
}

async fn get_with(state: Arc<AppState>, uri: &str) -> (StatusCode, serde_json::Value) {
    let res = crate::router(state)
        .oneshot(
            Request::builder()
                .uri(uri)
                .header("Authorization", "Bearer secret")
                .body(Body::empty())
                .unwrap(),
        )
        .await
        .unwrap();
    let status = res.status();
    let bytes = res.into_body().collect().await.unwrap().to_bytes();
    (
        status,
        serde_json::from_slice(&bytes).unwrap_or(serde_json::Value::Null),
    )
}

/// `/api/summary` returns the latest row of each host series in the shapes the
/// standalone `current` endpoints use, nested under cpu / memory / disk.
#[tokio::test]
async fn summary_returns_the_latest_of_each_series() {
    let store = Store::open_in_memory().unwrap();
    store.insert_cpu(1000, 10.0).unwrap();
    store.insert_cpu(2000, 42.5).unwrap(); // newer wins
    store.insert_memory(&mem_row(1000, 20)).unwrap();
    store.insert_memory(&mem_row(3000, 55)).unwrap(); // newer wins
    // Two mounts share the newest cycle timestamp; an older cycle must not leak.
    store.insert_disk_batch(500, &[disk("/", 5)]).unwrap();
    store
        .insert_disk_batch(4000, &[disk("/", 30), disk("/data", 70)])
        .unwrap();

    let (s, j) = get_with(state(store), "/api/summary").await;
    assert_eq!(s, StatusCode::OK);

    // cpu matches /api/cpu/current: string time, numeric percent.
    assert_eq!(j["cpu"]["time"], "2000");
    assert_eq!(j["cpu"]["percent"], 42.5);

    assert_eq!(j["memory"]["time"], "3000");
    assert_eq!(j["memory"]["used"], 55);
    assert_eq!(j["memory"]["usedPercent"], 55.0);

    let disks = j["disk"].as_array().unwrap();
    assert_eq!(disks.len(), 2, "only the newest cycle's mounts");
    assert_eq!(disks[0]["mount"], "/");
    assert_eq!(disks[0]["time"], "4000");
    assert_eq!(disks[1]["mount"], "/data");
    assert_eq!(disks[1]["used"], 70);
}

/// Every key is `null` when its table is empty.
#[tokio::test]
async fn summary_nulls_empty_series() {
    let (s, j) = get_with(state(Store::open_in_memory().unwrap()), "/api/summary").await;
    assert_eq!(s, StatusCode::OK);
    assert!(j["cpu"].is_null());
    assert!(j["memory"].is_null());
    assert!(j["disk"].is_null());
}

/// `/api/containers/current` returns one row per container carrying the latest
/// sample of each metric, keyed and joined by id.
#[tokio::test]
async fn containers_current_returns_latest_per_container() {
    let store = Store::open_in_memory().unwrap();
    // Two cycles: the second must win for both containers.
    store
        .insert_container_batch(
            1000,
            &[container("alpha", 5.0, 10), container("beta", 1.0, 2)],
        )
        .unwrap();
    store
        .insert_container_batch(
            2000,
            &[container("alpha", 88.0, 40), container("beta", 3.0, 6)],
        )
        .unwrap();
    // alpha also has a disk sample at a later time than its cpu/memory.
    store
        .insert_container_disk_batch(3000, &[container_disk("alpha", 12)])
        .unwrap();

    let (s, j) = get_with(state(store), "/api/containers/current").await;
    assert_eq!(s, StatusCode::OK);
    let rows = j.as_array().unwrap();
    assert_eq!(rows.len(), 2, "one row per container, sorted by id");

    let alpha = &rows[0];
    assert_eq!(alpha["id"], "alpha");
    // Container cpu reuses the history shape: percent is a string.
    assert_eq!(alpha["cpu"]["time"], "2000");
    assert_eq!(alpha["cpu"]["percent"], "88.00");
    assert_eq!(alpha["memory"]["time"], "2000");
    assert_eq!(alpha["memory"]["used"], 40);
    assert_eq!(alpha["disk"]["time"], "3000");
    assert_eq!(alpha["disk"]["writableLayer"], 12);
    assert_eq!(alpha["disk"]["volumesTotal"], 24);
    // Top-level time is the newest across the three (disk at 3000).
    assert_eq!(alpha["time"], "3000");

    let beta = &rows[1];
    assert_eq!(beta["id"], "beta");
    assert_eq!(beta["cpu"]["percent"], "3.00");
    // beta has no disk sample recorded.
    assert!(beta["disk"].is_null());
    assert_eq!(beta["time"], "2000");
}

/// A container with only a disk sample still appears, with cpu/memory `null`.
#[tokio::test]
async fn containers_current_includes_disk_only_container() {
    let store = Store::open_in_memory().unwrap();
    store
        .insert_container_disk_batch(1500, &[container_disk("lonely", 7)])
        .unwrap();

    let (s, j) = get_with(state(store), "/api/containers/current").await;
    assert_eq!(s, StatusCode::OK);
    let rows = j.as_array().unwrap();
    assert_eq!(rows.len(), 1);
    assert_eq!(rows[0]["id"], "lonely");
    assert!(rows[0]["cpu"].is_null());
    assert!(rows[0]["memory"].is_null());
    assert_eq!(rows[0]["disk"]["writableLayer"], 7);
    assert_eq!(rows[0]["time"], "1500");
}

/// An empty database yields an empty array, not an error.
#[tokio::test]
async fn containers_current_empty_db_is_empty_array() {
    let (s, j) = get_with(
        state(Store::open_in_memory().unwrap()),
        "/api/containers/current",
    )
    .await;
    assert_eq!(s, StatusCode::OK);
    assert_eq!(j, serde_json::json!([]));
}

/// Both endpoints sit under the global auth middleware: no bearer token → 401.
#[tokio::test]
async fn bulk_endpoints_require_auth() {
    for uri in [
        "/api/summary",
        "/api/containers/current",
        "/api/network/current",
        "/api/network/history",
        "/api/load/current",
        "/api/load/history",
    ] {
        let res = crate::router(state(Store::open_in_memory().unwrap()))
            .oneshot(Request::builder().uri(uri).body(Body::empty()).unwrap())
            .await
            .unwrap();
        assert_eq!(res.status(), StatusCode::UNAUTHORIZED, "{uri}");
    }
}

/// `/api/summary` carries the latest network rate, load average and host
/// status (uptime + swap) alongside cpu/memory/disk.
#[tokio::test]
async fn summary_includes_network_load_and_host() {
    let store = Store::open_in_memory().unwrap();
    store.insert_network(1000, 100.0, 50.0).unwrap();
    store.insert_network(2000, 123.5, 67.25).unwrap(); // newer wins
    store.insert_load(2000, 0.75, 1.5, 2.25).unwrap();
    store
        .upsert_host_status(&HostStatusRow {
            uptime_seconds: 86_400,
            swap_total: 2_000_000_000,
            swap_used: 500_000_000,
            swap_free: 1_500_000_000,
            swap_used_percent: 25.0,
            time: 2000,
        })
        .unwrap();

    let (s, j) = get_with(state(store), "/api/summary").await;
    assert_eq!(s, StatusCode::OK);
    assert_eq!(j["network"]["time"], "2000");
    assert_eq!(j["network"]["rxBytesPerSec"], 123.5);
    assert_eq!(j["network"]["txBytesPerSec"], 67.25);
    assert_eq!(j["load"]["load1"], 0.75);
    assert_eq!(j["load"]["load15"], 2.25);
    assert_eq!(j["host"]["uptimeSeconds"], 86_400);
    assert_eq!(j["host"]["swapUsedPercent"], 25.0);
    assert_eq!(j["host"]["swapUsed"], 500_000_000_u64);
}

/// The new host keys are `null` on an empty database, like the others.
#[tokio::test]
async fn summary_nulls_new_series_when_empty() {
    let (s, j) = get_with(state(Store::open_in_memory().unwrap()), "/api/summary").await;
    assert_eq!(s, StatusCode::OK);
    assert!(j["network"].is_null());
    assert!(j["load"].is_null());
    assert!(j["host"].is_null());
}

/// `/api/containers/current` rows carry the latest network rate and current
/// status; a container with only status (no metric samples) still appears.
#[tokio::test]
async fn containers_current_includes_network_and_status() {
    let store = Store::open_in_memory().unwrap();
    store
        .insert_container_batch(2000, &[container("alpha", 5.0, 10)])
        .unwrap();
    store
        .insert_container_network_batch(
            1000,
            &[ContainerNetworkSample {
                container_id: "alpha".into(),
                rx_bytes_per_sec: 10.0,
                tx_bytes_per_sec: 20.0,
            }],
        )
        .unwrap();
    store
        .insert_container_network_batch(
            3000,
            &[ContainerNetworkSample {
                container_id: "alpha".into(),
                rx_bytes_per_sec: 111.0,
                tx_bytes_per_sec: 222.0,
            }],
        )
        .unwrap();
    store
        .upsert_container_status_batch(
            2000,
            &[
                ContainerStatusSample {
                    container_id: "alpha".into(),
                    state: "running".into(),
                    health_status: "healthy".into(),
                    restart_count: 3,
                },
                // status-only container (never produced a metric sample).
                ContainerStatusSample {
                    container_id: "ghost".into(),
                    state: "exited".into(),
                    health_status: String::new(),
                    restart_count: 0,
                },
            ],
        )
        .unwrap();

    let (s, j) = get_with(state(store), "/api/containers/current").await;
    assert_eq!(s, StatusCode::OK);
    let rows = j.as_array().unwrap();
    assert_eq!(rows.len(), 2, "alpha + ghost");

    let alpha = &rows[0];
    assert_eq!(alpha["id"], "alpha");
    assert_eq!(alpha["network"]["time"], "3000", "latest rate wins");
    assert_eq!(alpha["network"]["rxBytesPerSec"], 111.0);
    assert_eq!(alpha["status"]["state"], "running");
    assert_eq!(alpha["status"]["health"], "healthy");
    assert_eq!(alpha["status"]["restartCount"], 3);
    // Top-level time is the newest metric sample (network 3000), not status.
    assert_eq!(alpha["time"], "3000");

    let ghost = &rows[1];
    assert_eq!(ghost["id"], "ghost");
    assert!(ghost["cpu"].is_null());
    assert!(ghost["network"].is_null());
    assert_eq!(ghost["status"]["state"], "exited");
    // No metric sample, so the top-level time falls back to 0.
    assert_eq!(ghost["time"], "0");
}

/// `/api/network/current` and `/api/load/current` return the latest row or
/// `null`; the history endpoints return time-ordered arrays.
#[tokio::test]
async fn network_and_load_endpoints() {
    let empty = Store::open_in_memory().unwrap();
    let (s, j) = get_with(state(empty), "/api/network/current").await;
    assert_eq!(s, StatusCode::OK);
    assert!(j.is_null(), "empty network is null");

    let store = Store::open_in_memory().unwrap();
    store.insert_network(1000, 10.0, 20.0).unwrap();
    store.insert_network(2000, 30.0, 40.0).unwrap();
    store.insert_load(1000, 0.1, 0.2, 0.3).unwrap();
    store.insert_load(2000, 0.4, 0.5, 0.6).unwrap();
    let st = state(store);

    let (s, j) = get_with(st.clone(), "/api/network/current").await;
    assert_eq!(s, StatusCode::OK);
    assert_eq!(j["rxBytesPerSec"], 30.0);
    assert_eq!(j["txBytesPerSec"], 40.0);

    let (s, j) = get_with(st.clone(), "/api/network/history").await;
    assert_eq!(s, StatusCode::OK);
    let rows = j.as_array().unwrap();
    assert_eq!(rows.len(), 2);
    assert_eq!(rows[0]["time"], "1000");
    assert_eq!(rows[1]["rxBytesPerSec"], 30.0);

    let (s, j) = get_with(st.clone(), "/api/load/current").await;
    assert_eq!(s, StatusCode::OK);
    assert_eq!(j["load1"], 0.4);

    let (s, j) = get_with(st, "/api/load/history").await;
    assert_eq!(s, StatusCode::OK);
    assert_eq!(j.as_array().unwrap().len(), 2);
}