# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentinel API",
    "description": "REST API for gathering Linux server and Docker Engine metrics.\n\nSentinel collects system metrics (CPU, memory) and Docker container statistics, storing them in SQLite and providing both current and historical data through REST endpoints.\n\nThis service is designed for integration with [Coolify.io](https://coolify.io).",
    "contact": {
      "name": "Coolify",
      "url": "https://coolify.io"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.2"
  },
  "paths": {
...
    "/api/containers/current": {
      "get": {
        "tags": [
          "Container Metrics"
        ],
        "summary": "Get all containers (current)",
        "description": "Latest CPU, memory, storage and network sample, plus container status, for every container in one response. Empty array when nothing recorded.",
        "operationId": "containers_current",
        "responses": {
          "200": {
            "description": "Current snapshot of every container",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContainerCurrent"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/cpu/current": {
      "get": {
        "tags": [
          "System Metrics"
        ],
        "summary": "Get current CPU usage",
        "description": "Retrieve the current CPU usage percentage",
        "operationId": "current",
        "responses": {
          "200": {
            "description": "Current CPU usage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CpuCurrent"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/cpu/history": {
      "get": {
        "tags": [
          "System Metrics"
        ],
        "summary": "Get CPU usage history",
        "description": "Retrieve historical CPU usage data with optional date range filtering",
        "operationId": "history",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Start date in ISO 8601 format (UTC). Defaults to the Unix epoch\n(all recorded history) when omitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End date in ISO 8601 format (UTC). Defaults to now when omitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical CPU usage data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CpuUsage"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/disk/current": {
      "get": {
        "tags": [
          "System Metrics"
        ],
        "summary": "Get current disk usage",
        "description": "Latest stored filesystem usage, one entry per real mountpoint",
        "operationId": "current",
        "responses": {
          "200": {
            "description": "Current disk usage per mountpoint",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiskUsage"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/disk/history": {
      "get": {
        "tags": [
          "System Metrics"
        ],
        "summary": "Get disk usage history",
        "description": "Historical filesystem usage across mountpoints with optional date range filtering",
        "operationId": "history",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Start date in ISO 8601 format (UTC). Defaults to the Unix epoch\n(all recorded history) when omitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End date in ISO 8601 format (UTC). Defaults to now when omitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical disk usage data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiskUsage"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/load/current": {
      "get": {
        "tags": [
          "System Metrics"
        ],
        "summary": "Get current load average",
        "description": "Latest host load average (1 / 5 / 15 minute), or null when nothing has been recorded yet",
        "operationId": "current",
        "responses": {
          "200": {
            "description": "Current load average",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/LoadAverage"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/load/history": {
      "get": {
        "tags": [
          "System Metrics"
        ],
        "summary": "Get load average history",
        "description": "Historical host load average with optional date range filtering",
        "operationId": "history",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Start date in ISO 8601 format (UTC). Defaults to the Unix epoch\n(all recorded history) when omitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End date in ISO 8601 format (UTC). Defaults to now when omitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Load average history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoadAverage"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/memory/current": {
      "get": {
        "tags": [
          "System Metrics"
        ],
        "summary": "Get current memory usage",
        "description": "Retrieve the current memory usage statistics",
        "operationId": "current",
        "responses": {
          "200": {
            "description": "Current memory usage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemUsage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/memory/history": {
      "get": {
        "tags": [
          "System Metrics"
        ],
        "summary": "Get memory usage history",
        "description": "Retrieve historical memory usage data with optional date range filtering",
        "operationId": "history",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Start date in ISO 8601 format (UTC). Defaults to the Unix epoch\n(all recorded history) when omitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End date in ISO 8601 format (UTC). Defaults to now when omitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical memory usage data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemUsage"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/network/current": {
      "get": {
        "tags": [
          "System Metrics"
        ],
        "summary": "Get current network usage",
        "description": "Latest host network throughput (bytes/sec, summed across physical interfaces), or null when nothing has been recorded yet",
        "operationId": "current",
        "responses": {
          "200": {
            "description": "Current network usage",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/NetworkUsage"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/network/history": {
      "get": {
        "tags": [
          "System Metrics"
        ],
        "summary": "Get network usage history",
        "description": "Historical host network throughput (bytes/sec) with optional date range filtering",
        "operationId": "history",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Start date in ISO 8601 format (UTC). Defaults to the Unix epoch\n(all recorded history) when omitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End date in ISO 8601 format (UTC). Defaults to now when omitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Network usage history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NetworkUsage"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/push-status": {
      "get": {
        "tags": [
          "Core"
        ],
        "summary": "Get push status",
        "description": "Outcome of the most recent pushes to Coolify: when the last attempt and the last success happened, the last error (DNS, TLS, connection, non-2xx response, ...) and how many attempts failed in a row. All fields are null / 0 when push is disabled or no push was attempted yet.",
        "operationId": "push_status",
        "responses": {
          "200": {
            "description": "Push status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushStatusBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/summary": {
      "get": {
        "tags": [
          "System Metrics"
        ],
        "summary": "Get host summary",
        "description": "Latest stored host CPU, memory, disk, network and load average usage, plus host status (uptime and swap), in one response. Any series with no rows is null.",
        "operationId": "summary",
        "responses": {
          "200": {
            "description": "Host summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostSummary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ContainerCurrent": {
        "type": "object",
        "description": "One container's latest snapshot for `GET /api/containers/current`. Each\nmetric reuses the same shape its per-container history endpoint returns and\nis `null` when that series has no rows for the container. `time` is the\nnewest millisecond timestamp (as a string, like every other `time`) across\nwhichever metric samples are present (`status` is a current-only row and\ndoes not affect it).",
        "required": [
          "id",
          "time"
        ],
        "properties": {
          "cpu": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CpuUsage"
              }
            ]
          },
          "disk": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ContainerDiskUsage"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "memory": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MemUsage"
              }
            ]
          },
          "network": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NetworkUsage"
              }
            ]
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ContainerStatus"
              }
            ]
          },
          "time": {
            "type": "string"
          }
        }
      },
...
      "CpuCurrent": {
        "type": "object",
        "description": "Current CPU usage. WIRE FORMAT IS FROZEN: `percent` is a NUMBER here,\nunlike /api/cpu/history where it is a string. Do not \"fix\" this.",
        "required": [
          "time",
          "percent"
        ],
        "properties": {
          "percent": {
            "type": "number",
            "format": "double"
          },
          "time": {
            "type": "string"
          }
        }
      },
      "CpuUsage": {
        "type": "object",
        "description": "WIRE FORMAT IS FROZEN. `percent` is a string here (history endpoints) but a\nnumber in /api/cpu/current. Do not \"fix\" this — Coolify parses it as-is.",
        "required": [
          "time",
          "percent"
        ],
        "properties": {
          "human_friendly_time": {
            "type": [
              "string",
              "null"
            ]
          },
          "percent": {
            "type": "string"
          },
          "time": {
            "type": "string"
          }
        }
      },
      "DiskUsage": {
        "type": "object",
        "description": "Server filesystem usage for one mountpoint. These endpoints are new (not\npart of the frozen Go wire format), so `time` is a stringified millis for\nconsistency with the other series while byte fields stay numeric.",
        "required": [
          "time",
          "mount",
          "total",
          "used",
          "available",
          "usedPercent"
        ],
        "properties": {
          "available": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "human_friendly_time": {
            "type": [
              "string",
              "null"
            ]
          },
          "mount": {
            "type": "string"
          },
          "time": {
            "type": "string"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "used": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "usedPercent": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "LoadAverage": {
        "type": "object",
        "description": "Host load average (1 / 5 / 15 minute).",
        "required": [
          "time",
          "load1",
          "load5",
          "load15"
        ],
        "properties": {
          "human_friendly_time": {
            "type": [
              "string",
              "null"
            ]
          },
          "load1": {
            "type": "number",
            "format": "double"
          },
          "load15": {
            "type": "number",
            "format": "double"
          },
          "load5": {
            "type": "number",
            "format": "double"
          },
          "time": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication. Set the TOKEN environment variable when running Sentinel and include it in the Authorization header."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}